# swagger.yaml
openapi: 3.0.0
info:
  title: API CertQuiz
  description: Endpoints para gerenciar o App CertQuiz
  version: 1.0.0
tags:
  - name: Provas
    description: Endpoints para gerenciar provas
  - name: Quiz
    description: Endpoints para gerenciar quiz
  - name: Questão
    description: Endpoints para gerenciar questões  
  - name: Gestão
    description: Endpoints para gestão do sistema
paths:
  /api/v1/login:
    post:
      summary: Realiza o login do usuário
      tags:
        - Gestão      
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                username:
                  type: string
                  description: Nome de usuário
                password:
                  type: string
                  description: Senha do usuário
              required:
                - username
                - password
      responses:
        '201':
          description: Login bem-sucedido
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: Token JWT para autenticação
                  expiresAt:
                    type: string
                    format: date-time
                    description: Data e hora de expiração do token
        '401':
          description: Credenciais inválidas
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: Mensagem de erro
  /api/v1/exam/{title}/quiz/{index}:
    get:
      summary: Obtém um quiz específico de um exame
      tags:
        - Quiz
      parameters:
        - name: title
          in: path
          required: true
          schema:
            type: string
          description: Título do exame
        - name: index
          in: path
          required: true
          schema:
            type: integer
          description: Índice do quiz
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
          description: Token JWT para autenticação
      responses:
        '200':
          description: Quiz retornado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  title:
                    type: string
                    description: Título do quiz
                  questions:
                    type: array
                    items:
                      type: object
                      properties:
                        question:
                          type: string
                          description: Pergunta do quiz
                        options:
                          type: array
                          items:
                            type: object
                            properties:
                              text:
                                type: string
                                description: Texto da opção
                              correct:
                                type: boolean
                                description: Indica se a opção está correta
        '401':
          description: Não autorizado
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: Mensagem de erro
  /api/v1/exams/:id:
    delete:
      summary: Deleta um exame pelo ID
      tags:
        - Provas
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: ID do recurso a ser deletado
      responses:
        200:
          description: Prova excluída com sucesso
        404:
          description: Prova não encontrada
        500:
          description: Erro ao excluir a prova
  /api/v1/exams:
    get:
      summary: Obtém a lista de exames
      tags: 
        - Provas
      parameters:
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
          description: Token JWT para autenticação
      responses:
        '200':
          description: Lista de exames retornada com sucesso
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    _id:
                      type: string
                      description: ID do exame
                    title:
                      type: string
                      description: Título do exame
                    quizzes:
                      type: array
                      items:
                        type: object
                        properties:
                          title:
                            type: string
                            description: Título do quiz
                          questions:
                            type: array
                            items:
                              type: object
                              properties:
                                question:
                                  type: string
                                  description: Pergunta do quiz
                                options:
                                  type: array
                                  items:
                                    type: object
                                    properties:
                                      text:
                                        type: string
                                        description: Texto da opção
                                      correct:
                                        type: boolean
                                        description: Indica se a opção está correta
        '401':
          description: Não autorizado
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: Mensagem de erro
    post:
      summary: Cria um novo exame
      tags:
        - Provas
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  description: Título do exame
                quizzes:
                  type: array
                  items:
                    type: object
                  description: Lista de quizzes do exame
              required:
                - title
                - quizzes
      responses:
        '201':
          description: Exame criado com sucesso
        '401':
          description: Não autorizado
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: Mensagem de erro
  /api/v1/question:
    post:
      summary: Adiciona uma nova questão a um quiz
      tags:
        - Questão
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                exam:
                  type: string
                  description: Nome do exame
                quiz:
                  type: string
                  description: Título do quiz
                newQuizTitle:
                  type: string
                  description: Novo título do quiz
                question:
                  type: string
                  description: Pergunta do quiz
                optionA:
                  type: string
                  description: Opção A
                optionB:
                  type: string
                  description: Opção B
                optionC:
                  type: string
                  description: Opção C
                optionD:
                  type: string
                  description: Opção D
                correctOption:
                  type: string
                  description: Opção correta
              required:
                - exam
                - quiz
                - question
                - optionA
                - optionB
                - optionC
                - optionD
                - correctOption
      parameters:
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
          description: Token JWT para autenticação
      responses:
        '200':
          description: Questão adicionada com sucesso
        '401':
          description: Não autorizado
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: Mensagem de erro
  /api/v1/quiz:
    post:
      summary: Adiciona um novo quiz a um exame
      tags:
        - Quiz
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                examTitle:
                  type: string
                  description: Título do exame
                quiz:
                  type: object
                  properties:
                    title:
                      type: string
                      description: Título do quiz
                    questions:
                      type: array
                      items:
                        type: object
                        properties:
                          question:
                            type: string
                            description: Pergunta do quiz
                          options:
                            type: array
                            items:
                              type: object
                              properties:
                                text:
                                  type: string
                                  description: Texto da opção
                                correct:
                                  type: boolean
                                  description: Indica se a opção está correta
              required:
                - examTitle
                - quiz
      parameters:
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
          description: Token JWT para autenticação
      responses:
        '200':
          description: Quiz adicionado com sucesso
        '401':
          description: Não autorizado
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: Mensagem de erro
  /api/v1/backup:
    get:
      summary: Realiza o backup dos dados
      tags:
        - Gestão
      parameters:
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
          description: Token JWT para autenticação
      responses:
        '200':
          description: Sucesso
          content:
            text/plain:
              schema:
                type: string
                example: Sucesso
        '401':
          description: Não autorizado
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: Mensagem de erro
  /api/v1/register:
    post:
      summary: Cria um novo usuário
      tags:
        - Gestão
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username:        
                  type: string
                password:  
                  type: string
              required:
                - username
                - password
      responses:
        201:
          description: Usuário registrado com sucesso
        500:
          description: Erro ao cadastrar o usuario